Move the matcher models out of internal_model and into the matcher
