+++ .sbt_metadata/merger.json
@@ -3,6 +3,7 @@
   "folder" : "pipeline/merger",
   "dependencyIds" : [
     "internal_model",
+    "matcher",
     "pipeline_storage_typesafe"
   ]
 }
