+++ pipeline/matcher/src/main/scala/weco/pipeline/matcher/storage/WorkNodeDao.scala
@@ -5,10 +5,10 @@ import org.scanamo.{DynamoFormat, Scanamo, Table}
 import org.scanamo.syntax._
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient
 import software.amazon.awssdk.services.dynamodb.model.ProvisionedThroughputExceededException
-import weco.catalogue.internal_model.matcher.WorkNode
-import weco.storage.dynamo.DynamoConfig
 import weco.catalogue.internal_model.identifiers.CanonicalId
+import weco.storage.dynamo.DynamoConfig
 import weco.pipeline.matcher.exceptions.MatcherException
+import weco.pipeline.matcher.models.WorkNode
 
 import scala.concurrent.{ExecutionContext, Future}
